model_pipeline/requirements.txt
boto3==1.26.138
botocore==1.29.138
click==8.1.3
imbalanced-learn==0.10.1
imblearn==0.0
iniconfig==2.0.0
jmespath==1.0.1
joblib==1.2.0
numpy==1.24.3
packaging==23.1
pandas==2.0.1
pluggy==1.0.0
pytest==7.3.1
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
s3transfer==0.6.1
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
threadpoolctl==3.1.0
typer==0.9.0
typing_extensions==4.6.2
tzdata==2023.3
urllib3==1.26.16
